GRN: fix PUT /garden 500 on background image upload
PUT /garden returned 500 with `Database query error: error serializing
parameter 4` whenever the frontend sent backgroundOpacity or
northOffsetDeg fields. The Rust model types those as i16 (matching the
smallint columns), but Postgres was inferring $5/$6 as integer from the
untyped literals in `coalesce($5, 60)` / `coalesce($6, 0)`. The
inferred-integer type didn't match the i16 the SDK was preparing to
send, so tokio_postgres rejected the call before it even reached
Postgres.

Fix: cast both the parameter and the literal to smallint explicitly
inside coalesce. The Rust signature stays as i16, so the column type
and the wire type now agree.

This also blocked the background-photo flow end-to-end — the upload to
S3 succeeded but the follow-up PUT /garden failed, so the canvas never
recorded the image key.

+++ services/grn-api/src/api/handlers/garden_canvas.rs
@@ -68,6 +68,10 @@ pub async fn update_my_canvas(
     validate_canvas_payload(&payload)?;
 
     let client = db::connect().await?;
+    // Explicit casts on $5/$6: Postgres otherwise infers their type as
+    // integer from the bare `60`/`0` literals in coalesce, but the Rust
+    // value is i16. Without the cast, tokio_postgres rejects the call
+    // with `error serializing parameter 4` (0-indexed → $5).
     let row = client
         .query_one(
             "
@@ -80,8 +84,8 @@ pub async fn update_my_canvas(
                 coalesce($2, 360),
                 coalesce($3, 240),
                 $4,
-                coalesce($5, 60),
-                coalesce($6, 0)
+                coalesce($5::smallint, 60::smallint),
+                coalesce($6::smallint, 0::smallint)
             )
             on conflict (user_id) do update
               set width_inches = coalesce(excluded.width_inches, garden_canvases.width_inches),
